浣熊/pyproject.toml
[project]
name = "raccoon"
version = "0.1.0"
description = "浣熊 - LLM API Testing Platform / 大模型API测试平台"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0",
    "anthropic>=0.30",
    "httpx>=0.25",
    "pyyaml>=6.0",
    "typer>=0.12",
    "rich>=13.0",
    "fastapi>=0.110",
    "uvicorn>=0.29",
    "jinja2>=3.1",
    "apscheduler>=3.10,<4",
]

[project.scripts]
raccoon = "raccoon.cli:app"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
